Moved frontend reCAPTCHA logic to util file
